+++ src/index.js
@@ -1,4 +1,13 @@
 export { resolveEffectiveView } from "./resolve.js";
+export {
+  REPORT_ONLY_EVALUATION_READBACK_VERSION,
+  REPORT_ONLY_EVALUATION_RECEIPT_VERSION,
+  REPORT_ONLY_PROOF_RUNG,
+  createReportOnlyEvaluationReadback,
+  hashReportOnlyEvaluationReceipt,
+  resolveReportOnlyEvaluationReceipt,
+  verifyReportOnlyEvaluationReadback
+} from "./evaluation-receipt.js";
 export { resolvePolicyPackView } from "./policy-pack.js";
 export { buildResolutionContext, matchesPattern, matchesScope, matchesWhen } from "./match.js";
 export { collectRuleEntries } from "./cascade.js";